Tweaked new verifier APIs according to the design doc.
Test: manual testing, unit tests.
