+++ multipaz/src/commonMain/kotlin/org/multipaz/verification/DcqlRequestDefinition.kt
@@ -3,7 +3,7 @@ package org.multipaz.verification
 /**
  * [RequestDefinition] that is defined in terms of OpenID4VP DCQL and transaction data.
  *
- * @param dcql DCQL query as parsed JSON, this gives the list of needed credentials and
+ * @property dcql DCQL query, this gives the list of needed credentials and
  *   which claims are needed from each credential
  * @property transactionData transaction data in OpenID4VP JSON format (before Base64Url
  *   encoding), note that credentialId uses credential ids used in DCQL
